chore(pom.xml): add maven-javadoc-plugin to generate API documentation for the project
The maven-javadoc-plugin is added to the project's pom.xml file. This plugin is used to generate API documentation for the project. The configuration includes setting the source version to 1.8, disabling author information, setting the doctitle, enabling quiet mode, disabling doclint, setting the encoding to UTF-8, and specifying the output directory for the generated JAR file. The plugin is executed during the package phase.